requirements.txt
sentence-transformers==2.6.1
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-groq
streamlit
-e .